+++ lib/src/clojure_lsp/db.clj
@@ -12,6 +12,7 @@
 (defonce db (atom {:documents {}}))
 (defonce current-changes-chan (async/chan 1))
 (defonce diagnostics-chan (async/chan 1))
+(defonce created-watched-files-chan (async/chan 1))
 (defonce edits-chan (async/chan 1))
 
 (def version 1)